requirements.txt
pretrainedmodels
pytorch0.3.0,
tensorboard_logger,
pretrainedmodels,
argparse,
PIL